CMake!
project now is CMake-able

Moved glm glfw to git submodules

added tiny-obj header

fixed some safety bugs

oh, and fixed the damn CMakeLists.txt

+++ Vulkan Engine/renderer/simple_render_system/vk_descriptors.cpp
@@ -201,5 +201,5 @@ void vk_descriptor_writer::overwrite(const VkDescriptorSet& set)
 	for (auto& write : writes)
 		write.dstSet = set;
 
-	vkUpdateDescriptorSets(pool.device.get_device(), writes.size(), writes.data(), 0, nullptr);
+	vkUpdateDescriptorSets(pool.device.get_device(), static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
 }
